Create miscFuns.py
This is a module of helpful functions for use in our project. Import this module to access these functions like you would any other module/library (i.e., "import miscFuns as mf" or something like that). Consider this a dumping ground for functions that feel like they detract from the flow of the notebook or are useful to share across multiple files/users.