ADSR v1.1 and miscFuns updates
Fixed bug in ADSR where the segmentation points were continuously redefined. Changed parameter values, but I also found that the parameter values are a bug in and of themselves. In brief, the parameter values (sound threshold, steady-state slope, and dx of average slope calculation) need to be changed for each sound in order to pick up the sound's ADSR values properly. For now, we will keep one set of parameters and hope it works out ok.

calcADSR and envelope functions were copied over into miscFuns so that they can be used without importing the Feature Vectors.ipynb file, which will remain a file for demos and experimentation.