Added Beatboxing_Detection.py and model2csv/csv2model
No longer have to re-generate the model from scratch, general model can be constructed from csv files. Beatboxing_Detection.py is now considered the official file to use as an import, along with miscFuns.py. The Jupyter Notebooks will remain as a way to experiment and develop, but they should not be part of official releases (unless they are used as documentation)